Updated the Compression filter's integration tests to extract the test for ensuring the compressed data was written to the stream and the Content-Length header was updated accordingly.
