+++ repose-aggregator/tests/functional-tests/src/integrationTest/groovy/features/filters/compression/CompressionHeaderTest.groovy
@@ -311,43 +311,4 @@ class CompressionHeaderTest extends ReposeValveTest {
         "Accept-Encoding" | "deflate"
         "accept-Encoding" | "identity"
     }
-
-    @Unroll("A GET request with Accept-Encoding header set to #encoding is honored on the output")
-    def "Check if GET request with Accept-Encoding header set to #encoding is honored on the output"() {
-        when:
-        "the content is sent to the origin service through Repose with Accept-Encoding " + encoding
-        def MessageChain mc = deproxy.makeRequest(
-                url: reposeEndpoint,
-                method: 'GET',
-                headers: [
-                        'Content-Length' : '0',
-                        'Accept-Encoding': encoding
-                ],
-                defaultHandler: { new Response(HttpStatus.SC_OK, "OK", null, content) }
-        )
-
-        then: "the uncompressed response from the origin service should be compressed before reaching the client"
-        mc.handlings[0].response.body.toString() == content
-        if (mc.receivedResponse.body instanceof byte[]) {
-            // This doesn't seem to work because of extra -1's padding the end of the body array.
-            //Arrays.equals((byte[]) mc.receivedResponse.body, zippedContent)
-            // So we go old school.
-            def rxBody = (byte[]) mc.receivedResponse.body
-            def expected = (byte[]) zippedContent
-            for (int i = 0; i < expected.length; i++) {
-                assert (rxBody[i] == expected[i])
-            }
-        } else {
-            assert (zippedContent instanceof String)
-            assert (mc.receivedResponse.body == zippedContent)
-        }
-        Integer.valueOf(mc.receivedResponse.headers.getFirstValue(HttpHeaders.CONTENT_LENGTH, "-1")) == zippedContent.length
-
-        where:
-        encoding   | zippedContent
-        "gzip"     | gzipCompressedContent
-        "x-gzip"   | gzipCompressedContent
-        "deflate"  | deflateCompressedContent
-        "identity" | content
-    }
 }
